requirements.txt
# Dépendances Python pour le projet IoT/Blockchain

# MQTT
paho-mqtt>=2.0.0

# Email
# (Utilise les modules intégrés: imaplib, smtplib, email)

# Hedera Hashgraph SDK
hedera-sdk-py>=2.0.0

# Variables d'environnement
python-dotenv>=1.0.0

# Interface Dashboard
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0
streamlit-autorefresh
scikit-learn



# Base de données
# SQLite est intégré à Python, pas besoin de dépendance

# Tests
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilitaires
requests>=2.31.0
fpdf
